feat: add mutation observe as a property of player
Player as a ff had to be fine-tuned to its proper role in the app.
Besides not having even a real reason to exist prior to the changes, the
mutation observer related properties fit just nice in what you would
expect from the a "Player" ff in this context, even semantically.