feat: and refactor. Modularizes responsibilities.
Although barebones, the previous iteration of the module was heading in
a direction in which it was gonna do too many things, so it was dialled
back to just work as a trigger(setObservers, setBoard) and a really
simple logic module(changeActivePlayer, endGame).
The intent is that driver only has to get the ball rolling and help
in menial tasks, as the data is stored on the gameboards(players) and
the html interface dictates its interactivity with css's help.