feat: randomly place ships at match start.
For the auto placement of ships to work, placeShip() had to be improved
to check if the available requested coordinate was available.
Two more things had to be determined: what to make of the coordinates
given (they were determined as the starting point of the ship) and
how to determine orientation (parameter of the same name receives
argument with 0 or 1 value, 0 indicating horizontal and 1 vertical).